requirements.txt
pandas>=2.0.0
polars>=1.5.0
numpy>=1.25.0
scipy>=1.11.0
yfinance>=0.2.43
requests>=2.31.0
python-dotenv>=1.0.1
pydantic>=2.8.2
pydantic-settings>=2.0.0
SQLAlchemy>=2.0.30
duckdb>=1.0.0
duckdb-engine>=0.11.8
mlflow>=2.14.0
scikit-learn>=1.4.2
xgboost>=2.0.3
vectorbt>=0.26.7
prefect>=2.19.0
ib-insync>=0.9.86
typer>=0.12.3
streamlit>=1.38.0
plotly>=5.22.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
twilio>=9.1.0
python-multipart>=0.0.9
pytest>=7.4.0
openpyxl>=3.1.0
reportlab>=4.0.0
schedule>=1.2.0
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
psutil>=5.9.0
slowapi>=0.1.9
redis>=5.0.0
sentry-sdk>=1.40.0
websockets>=12.0
alembic>=1.13.0
feedparser>=6.0.10
vaderSentiment>=3.3.2
polygon-api-client>=1.13.0
alpaca-trade-api>=3.1.1